@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <Version>4.1.7</Version>
     <Description>The storefront implementation of the Virto Commerce platform.</Description>
     <Copyright>Copyright © 2011-2019 Virto Commerce - an enterprise Microsoft ecommerce platform ©. All rights reserved</Copyright>
@@ -18,11 +18,11 @@
 
 
   <ItemGroup>
+	<FrameworkReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="LibSassHost" Version="1.2.10" />
     <PackageReference Include="LibSassHost.Native.linux-x64" Version="1.2.10" />
     <PackageReference Include="LibSassHost.Native.win-x64" Version="1.2.10" />
     <PackageReference Include="LibSassHost.Native.win-x86" Version="1.2.10" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="Nito.AsyncEx" Version="5.0.0" />
     <PackageReference Include="PagedList.Core" Version="1.17.4" />
     <PackageReference Include="Scriban" Version="2.1.1" />
