@@ -1,11 +1,10 @@
 using System.IO;
 using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
-using System.Threading.Tasks;
 using VirtoCommerce.LiquidThemeEngine;
-using VirtoCommerce.Storefront.Common;
 using VirtoCommerce.Storefront.Extensions;
 using VirtoCommerce.Storefront.Infrastructure;
 using VirtoCommerce.Storefront.Model;
@@ -20,10 +19,10 @@ namespace VirtoCommerce.Storefront.Controllers
     {
         private readonly ILiquidThemeEngine _themeEngine;
         private readonly IContentBlobProvider _contentBlobProvider;
-        private readonly IHostingEnvironment _hostingEnvironment;
+        private readonly IWebHostEnvironment _hostingEnvironment;
 
         public AssetController(IWorkContextAccessor workContextAccessor, IStorefrontUrlBuilder urlBuilder, ILiquidThemeEngine themeEngine,
-                               IContentBlobProvider staticContentBlobProvider, IHostingEnvironment hostingEnvironment)
+                               IContentBlobProvider staticContentBlobProvider, IWebHostEnvironment hostingEnvironment)
             : base(workContextAccessor, urlBuilder)
         {
             _themeEngine = themeEngine;
