@@ -1,20 +1,20 @@
 using System.Linq;
 using Microsoft.AspNetCore.Authorization;
-using Swashbuckle.AspNetCore.Swagger;
+using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
 namespace VirtoCommerce.Storefront.Infrastructure.Swagger
 {
     public class AuthResponsesOperationFilter : IOperationFilter
     {
-        public void Apply(Operation operation, OperationFilterContext context)
+        public void Apply(OpenApiOperation operation, OperationFilterContext context)
         {
             var authAttributes = context.MethodInfo.DeclaringType.GetCustomAttributes(true)
                 .Union(context.MethodInfo.GetCustomAttributes(true))
                 .OfType<AuthorizeAttribute>();
 
             if (authAttributes.Any())
-                operation.Responses.Add("401", new Response { Description = "Unauthorized" });
+                operation.Responses.Add("401", new OpenApiResponse { Description = "Unauthorized" });
         }
     }
 }
